# frontend/kucukdevapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "kucukdevapi",
    "version": "1.0.0"
  },
  "paths": {
...
    "/users/{uid}/semesters": {
      "get": {
        "tags": [
          "semesters"
        ],
        "summary": "List Semesters",
        "description": "list all semesters of a user with given userID",
        "operationId": "listSemestersOfUser",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            },
            "name": "uid",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "List all semesters",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Semesters Users  Uid  Semesters Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SemesterAPIModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "semesters"
        ],
        "summary": "Create Semester",
        "description": "Create a semester for a user with given userID",
        "operationId": "createSemester",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            },
            "name": "uid",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSemesterModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Add new semester",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCreate"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCreate"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/users/{uid}/semesters/{sid}": {
      "get": {
        "tags": [
          "semesters"
        ],
        "summary": "Show Semester",
        "description": "Get a single semester with given userID and semesterID",
        "operationId": "getSingleSemester",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            },
            "name": "uid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            },
            "name": "sid",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Get a single semester",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemesterAPIModel"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "semesters"
        ],
        "summary": "Update Semester",
        "description": "Update a semester with given userID and semesterID",
        "operationId": "updateSemester",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            },
            "name": "uid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            },
            "name": "sid",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserSemesterModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update a semester",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "semesters"
        ],
        "summary": "Delete Semester",
        "description": "Delete a semester with given userID and semesterID",
        "operationId": "deleteSemester",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            },
            "name": "uid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            },
            "name": "sid",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete semester",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbsenceModel": {
        "title": "AbsenceModel",
        "required": [
          "week",
          "day",
          "hour",
          "isLab"
        ],
        "type": "object",
        "properties": {
          "week": {
            "title": "Week",
            "minimum": 0,
            "type": "integer"
          },
          "day": {
            "title": "Day",
            "maximum": 4,
            "minimum": 0,
            "type": "integer"
          },
          "hour": {
            "title": "Hour",
            "maximum": 15,
            "minimum": 0,
            "type": "integer"
          },
          "isLab": {
            "title": "Islab",
            "maximum": 1,
            "minimum": 0,
            "type": "integer"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "LessonModel": {
        "title": "LessonModel",
        "required": [
          "name",
          "instructor",
          "absenceLimit",
          "slots"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "minLength": 1,
            "type": "string"
          },
          "instructor": {
            "title": "Instructor",
            "minLength": 1,
            "type": "string"
          },
          "absenceLimit": {
            "title": "Absencelimit",
            "minimum": 0,
            "type": "integer"
          },
          "slots": {
            "title": "Slots",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SlotModel"
            }
          },
          "absences": {
            "title": "Absences",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AbsenceModel"
            },
            "default": []
          }
        },
        "example": {
          "name": "Algebra",
          "instructor": "Jack Joe",
          "absenceLimit": 21,
          "slots": [
            {
              "day": 2,
              "hour": 7,
              "isLab": 0
            },
            {
              "day": 2,
              "hour": 8,
              "isLab": 0
            }
          ],
          "absences": [
            {
              "week": 0,
              "day": 2,
              "hour": 7,
              "isLab": 0
            },
            {
              "week": 0,
              "day": 2,
              "hour": 8,
              "isLab": 0
            }
          ]
        }
      },
      "Message": {
        "title": "Message",
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        }
      },
      "MessageCreate": {
        "title": "MessageCreate",
        "required": [
          "_id",
          "message"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "example": {
          "_id": "61ddea901311ecaed99afb7c",
          "message": "string"
        }
      },
      "SemesterAPIModel": {
        "title": "SemesterAPIModel",
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "startDate": {
            "title": "Startdate",
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "title": "Enddate",
            "type": "string",
            "format": "date-time"
          },
          "startHour": {
            "title": "Starthour",
            "type": "string"
          },
          "dLesson": {
            "title": "Dlesson",
            "type": "integer"
          },
          "dBreak": {
            "title": "Dbreak",
            "type": "integer"
          },
          "slotCount": {
            "title": "Slotcount",
            "type": "integer"
          }
        },
        "example": {
          "_id": "61ddea901311ecaed99afb7c",
          "name": "2020-21 Spring",
          "startDate": "2022-02-18T00:00:00Z",
          "endDate": "2022-06-18T00:00:00Z",
          "startHour": "8.10",
          "dLesson": 50,
          "dBreak": 10,
          "slotCount": 12
        }
      },
      "SlotModel": {
        "title": "SlotModel",
        "required": [
          "day",
          "hour",
          "isLab"
        ],
        "type": "object",
        "properties": {
          "day": {
            "title": "Day",
            "maximum": 4,
            "minimum": 0,
            "type": "integer"
          },
          "hour": {
            "title": "Hour",
            "maximum": 15,
            "minimum": 0,
            "type": "integer"
          },
          "isLab": {
            "title": "Islab",
            "maximum": 1,
            "minimum": 0,
            "type": "integer"
          }
        }
      },
...
      "UpdateUserSemesterModel": {
        "title": "UpdateUserSemesterModel",
        "required": [
          "name",
          "startDate",
          "endDate",
          "startHour",
          "dLesson",
          "dBreak",
          "slotCount"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "startDate": {
            "title": "Startdate",
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "title": "Enddate",
            "type": "string",
            "format": "date-time"
          },
          "startHour": {
            "title": "Starthour",
            "type": "string"
          },
          "dLesson": {
            "title": "Dlesson",
            "exclusiveMinimum": 0,
            "type": "integer"
          },
          "dBreak": {
            "title": "Dbreak",
            "exclusiveMinimum": 0,
            "type": "integer"
          },
          "slotCount": {
            "title": "Slotcount",
            "exclusiveMaximum": 16,
            "exclusiveMinimum": 3,
            "type": "integer"
          }
        },
        "example": {
          "name": "2020-21 Spring",
          "startDate": "2022-02-18T00:00:00Z",
          "endDate": "2022-06-18T00:00:00Z",
          "startHour": "8.10",
          "dLesson": 50,
          "dBreak": 10,
          "slotCount": 12
        }
      },
...
      "UserSemesterModel": {
        "title": "UserSemesterModel",
        "required": [
          "name",
          "startDate",
          "endDate",
          "startHour",
          "dLesson",
          "dBreak",
          "slotCount"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "startDate": {
            "title": "Startdate",
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "title": "Enddate",
            "type": "string",
            "format": "date-time"
          },
          "startHour": {
            "title": "Starthour",
            "type": "string"
          },
          "dLesson": {
            "title": "Dlesson",
            "exclusiveMinimum": 0,
            "type": "integer"
          },
          "dBreak": {
            "title": "Dbreak",
            "exclusiveMinimum": 0,
            "type": "integer"
          },
          "slotCount": {
            "title": "Slotcount",
            "exclusiveMaximum": 16,
            "exclusiveMinimum": 3,
            "type": "integer"
          },
          "lessons": {
            "title": "Lessons",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonModel"
            },
            "default": []
          }
        },
        "example": {
          "name": "2020-21 Spring",
          "startDate": "2022-02-18T00:00:00Z",
          "endDate": "2022-06-18T00:00:00Z",
          "startHour": "8.10",
          "dLesson": 50,
          "dBreak": 10,
          "slotCount": 12
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}